Added player model and camera lerping

+++ src/Utils/Time.java
@@ -4,7 +4,7 @@ import static org.lwjgl.glfw.GLFW.glfwGetTime;
 
 public class Time {
 
-    public static double deltaTime = 0.0d;
+    public static float deltaTime = 0.0f;
     public static double getTimeSeconds() { return glfwGetTime(); }
 
 }
